Move cache to the "util" module
